+++ include/common.hpp
@@ -4,9 +4,12 @@
  */
 #pragma once
 #define MATH_LIB_VER_MAJOR 0
-#define MATH_LIB_VER_MINOR 10
+#define MATH_LIB_VER_MINOR 11
 #define MATH_LIB_VER_PATCH 0
 #define MATH_LIB_VER_ALPHA 0 // 0 if release version, 1 if alpha version
 #define MATH_LIB_ENABLE_CANARY_MODE false
 #define MATH_LIB_INLINE_AGGRESSIVELY true
 
+#define PROCESSOR_TYPE_OTHER 0
+#define PROCESSOR_TYPE_TMS320C6748 1
+#define PROCESSOR_TYPE PROCESSOR_TYPE_OTHER